cleaned up code
