+++ ControlStack/CMakeLists.txt
@@ -45,7 +45,7 @@ find_library(MUJOCO_LIB mujoco REQUIRED HINTS ${MUJOCO_DIR}/lib)
 add_library(mujoco SHARED IMPORTED GLOBAL)
 set_property(TARGET mujoco PROPERTY IMPORTED_LOCATION ${CMAKE_INCLUDE_DIR}/lib/libmujoco.so)
 # add the executable
-add_executable(hopper_ctrl src/Controller.cpp src/Hopper.cpp src/Policy.cpp src/ZeroDynamicsPolicy.cpp)
+add_executable(hopper_ctrl src/Controller.cpp src/Hopper.cpp src/Policy.cpp)
 target_link_libraries(hopper_ctrl yaml-cpp)
 target_link_libraries(hopper_ctrl Eigen3::Eigen)
 target_link_libraries(hopper_ctrl onnxruntime::onnxruntime)
@@ -59,7 +59,7 @@ message(HARDWARE_STATUS="${HARDWARE}")
 
 if(HARDWARE)
     message(STATUS "Hardware is building")
-    add_executable(hopper_ctrl_hardware src/Controller_Hardware.cpp src/Hopper.cpp src/Policy.cpp src/ZeroDynamicsPolicy.cpp)
+    add_executable(hopper_ctrl_hardware src/Controller_Hardware.cpp src/Hopper.cpp src/Policy.cpp)
     # target_link_libraries(hopper_ctrl_hardware MPC)
     target_link_libraries(hopper_ctrl_hardware yaml-cpp)
     target_link_libraries(hopper_ctrl_hardware Eigen3::Eigen)
